+++ src/Dime/FrontendBundle/Resources/public/lib/src/component/edit/offer_edit.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:html';
 
-import 'package:DimeClient/src/component/elements/dime_form_group.dart';
+import 'package:DimeClient/src/component/elements/validation/dime_form_group.dart';
 import 'package:angular/angular.dart';
 import 'package:angular_forms/angular_forms.dart';
 import 'package:angular_router/angular_router.dart';
