+++ src/Dime/FrontendBundle/Resources/public/lib/src/component/elements/dime_directives.dart
@@ -4,7 +4,9 @@ import 'dime-button.dart';
 import 'download-button.dart';
 import 'error_icon.dart';
 import 'help-tooltip.dart';
-import 'dime_form_group.dart';
+import 'validation/dime_form_group.dart';
+import 'validation/dime_box.dart';
+import 'validation/validation_status_directive.dart';
 import 'discount_input.dart';
 import 'save_button.dart';
 import 'percentage_input.dart';
